fix user_id does not have a default value error when adding book

+++ routes/web.php
@@ -19,18 +19,18 @@ use App\Http\Controllers\UserController;
 Route::get("/register", [UserController::class, "register"])->name("users.register");
 Route::post("/register", [UserController::class, "store"]);
 
-Route::get("/login", [UserController::class, "login"])->name("login")->name("users.login");
+Route::get("/login", [UserController::class, "login"])->name("login");
 Route::post("/login", [UserController::class, "authenticate"]);
 
-Route::post("/logout", [UserController::class, "logout"])->name("logout")->name("users.logout");
+Route::post("/logout", [UserController::class, "logout"])->name("logout")->name("users.logout")->middleware("auth");
+Route::get("/addBook", [BookController::class, "addBook"])->name("books.add")->middleware("auth");
+Route::post("/addBook", [BookController::class, "storeBook"])->middleware("auth");
+
+Route::get("{book}/edit", [BookController::class, "editBook"])->name("books.edit")->middleware("auth");
+Route::put("{book}/edit", [BookController::class, "updateBook"])->middleware("auth");
+Route::delete("{book}/delete", [BookController::class, "deleteBook"])->name("books.delete")->middleware("auth");
 
 Route::get("/", [BookController::class, "allBooks"])->name("users.index");
-
-Route::get("/addBook", [BookController::class, "addBook"])->name("books.add");
-Route::post("/addBook", [BookController::class, "storeBook"]);
 Route::get("/{book}", [BookController::class, "singleBook"])->name("books.book");
 
-Route::get("{book}/edit", [BookController::class, "editBook"])->name("books.edit");
-Route::put("{book}/edit", [BookController::class, "updateBook"])->name("books.update");
-Route::delete("{book}/delete", [BookController::class, "deleteBook"])->name("books.delete");
